backend/requirements.txt
uvicorn
fastapi
strawberry-graphql[fastapi]
psycopg2-binary
SQLAlchemy
python-dotenv
pydantic
pydantic[email]
passlib[bcrypt]
bcrypt
Pillow
PyJWT
python-multipart
websockets
minio
aiodataloader
